Now we can add more options to the dropdow print button.
